split up in extra components
